+++ spec/factories/articles.rb
@@ -1,7 +1,7 @@
 FactoryGirl.define do
   factory :article do
-    sequence(:title) {|n| "Article #{n}"}
-    sequence(:slug) { |n| Faker::Lorem.words(n, true).join('-') }
+    sequence(:title) { |n| "Article #{n}" }
+    sequence(:slug) { |n| "article-#{n}-slug-#{Time.now.to_i}" }
     content do
       content = ''
       3.times do